+++ packages/autorest.python/test/multiapi/Expected/AcceptanceTests/Multiapi/multiapi/v1/_vendor.py
@@ -34,3 +34,9 @@ class MultiapiServiceClientMixinABC(ABC):
     _config: MultiapiServiceClientConfiguration
     _serialize: "Serializer"
     _deserialize: "Deserializer"
+
+
+def _curly_braces_encode(http_url: str) -> str:
+    for item in ("{", "}"):
+        http_url = http_url.replace(item, item * 2)
+    return http_url